another typo in vehicle api, refs #2942
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@23505 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/traci-server/TraCIServerAPI_Vehicle.cpp
@@ -422,7 +422,7 @@ TraCIServerAPI_Vehicle::processGet(TraCIServer& server, tcpip::Storage& inputSto
                 break;
             case VAR_SPEEDSETMODE:
                 tempMsg.writeUnsignedByte(TYPE_INTEGER);
-                tempMsg.writeDouble(TraCI_Vehicle::getSpeedMode(id));
+                tempMsg.writeInt(TraCI_Vehicle::getSpeedMode(id));
                 break;
             case VAR_LINE:
                 tempMsg.writeUnsignedByte(TYPE_STRING);
